moving the data containers to containers.py
